Factor out MultiEnumSearchContext from MultiNumericEnumSearchContext.
