+++ searchlib/src/vespa/searchlib/attribute/multi_numeric_enum_search_context.hpp
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "multi_numeric_enum_search_context.h"
+#include "multi_enum_search_context.hpp"
 #include "attributeiterators.hpp"
 #include <vespa/searchlib/queryeval/emptysearch.h>
 
@@ -10,10 +11,7 @@ namespace search::attribute {
 
 template <typename T, typename M>
 MultiNumericEnumSearchContext<T, M>::MultiNumericEnumSearchContext(std::unique_ptr<QueryTermSimple> qTerm, const AttributeVector& toBeSearched, const MultiValueMapping<M>& mv_mapping, const EnumStoreT<T>& enum_store)
-    : NumericRangeMatcher<T>(*qTerm),
-      SearchContext(toBeSearched),
-      _mv_mapping(mv_mapping),
-      _enum_store(enum_store)
+    : MultiEnumSearchContext<T, NumericRangeMatcher<T>, M>(NumericRangeMatcher<T>(*qTerm), toBeSearched, mv_mapping, enum_store)
 {
 }
 
@@ -28,10 +26,10 @@ template <typename T, typename M>
 std::unique_ptr<queryeval::SearchIterator>
 MultiNumericEnumSearchContext<T, M>::createFilterIterator(fef::TermFieldMatchData* matchData, bool strict)
 {
-    if (!valid()) {
+    if (!this->valid()) {
         return std::make_unique<queryeval::EmptySearch>();
     }
-    if (getIsFilter()) {
+    if (this->getIsFilter()) {
         return strict
             ? std::make_unique<FilterAttributeIteratorStrict<MultiNumericEnumSearchContext>>(*this, matchData)
             : std::make_unique<FilterAttributeIteratorT<MultiNumericEnumSearchContext>>(*this, matchData);
